Solving the growing to top problem the best way
During the development, when the user clicked the button, the main div was growing to the top and botton, when the objective was growing just to the botton. So, my solution was defining a height to the main div, and when the paragraf shows, questions container growed to the botton, keeping the main size static. The funcionality was working well, but was not ideal. So, a defined a top 70px and position absolute to the main div, preventing that it growed to the top. This way I could build a better logic and sintax to the code